Move graphical widget rendering to theme specific modules
