--- backend-embedded-graphics/src/widgets/graphical/mod.rs
@@ -1,2 +1,0 @@
-pub mod checkbox;
-pub mod radio;
