+++ backend-embedded-graphics/src/widgets/mod.rs
@@ -1,6 +1,5 @@
 pub mod background;
 pub mod border;
-pub mod graphical;
 pub mod label;
 pub mod scroll;
 pub mod text_block;
